# Conflicts:
#	ballin/main.cpp

+++ ballin/Shader.cpp
@@ -146,9 +146,9 @@ unsigned char* Shader::readShaderFile(const char *filename) {
         printError("ERROR", "Cannot open shader file!");
   		  return 0;
     }
-    int bytesinfile = filelength(file);
+    int bytesinfile = (int)filelength(file);
     unsigned char *buffer = new unsigned char[bytesinfile+1];
-    int bytesread = fread( buffer, 1, bytesinfile, file);
+    int bytesread = (int)fread( buffer, 1, bytesinfile, file);
     buffer[bytesread] = 0; // Terminate the string with 0
     fclose(file);
